[BEAM-1218] Move the dataflow specific parts from runner_test to dataflow_runner_test
